requirements.txt
fastapi>=0.109.1
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.4.0
requests
